+++ MoMoney.Core/ViewModels/Settings/BulkEditingViewModel.cs
@@ -1,7 +1,7 @@
 ﻿using System.Text;
+using System.Collections.ObjectModel;
 using CommunityToolkit.Mvvm.Input;
 using CommunityToolkit.Mvvm.ComponentModel;
-using MvvmHelpers;
 using UraniumUI.Material.Controls;
 using MoMoney.Core.Models;
 using MoMoney.Core.Helpers;
@@ -10,27 +10,27 @@ using MoMoney.Core.Services.Interfaces;
 
 namespace MoMoney.Core.ViewModels.Settings;
 
-public partial class BulkEditingViewModel : CommunityToolkit.Mvvm.ComponentModel.ObservableObject
+public partial class BulkEditingViewModel : ObservableObject
 {
     readonly IAccountService accountService;
     readonly ICategoryService categoryService;
     readonly ITransactionService transactionService;
     readonly ILoggerService<BulkEditingViewModel> logger;
 
-    [ObservableProperty] ObservableRangeCollection<Account> accounts = [];
+    [ObservableProperty] ObservableCollection<Account> accounts = [];
     [ObservableProperty] Account findAccount;
     [ObservableProperty] Account replaceAccount;
 
-    [ObservableProperty] ObservableRangeCollection<Category> categories = [];
+    [ObservableProperty] ObservableCollection<Category> categories = [];
     [ObservableProperty] Category findCategory;
     [ObservableProperty] Category replaceCategory;
 
-    [ObservableProperty] ObservableRangeCollection<Category> findSubcategories = [];
-    [ObservableProperty] ObservableRangeCollection<Category> replaceSubcategories = [];
+    [ObservableProperty] ObservableCollection<Category> findSubcategories = [];
+    [ObservableProperty] ObservableCollection<Category> replaceSubcategories = [];
     [ObservableProperty] Category findSubcategory;
     [ObservableProperty] Category replaceSubcategory;
 
-    [ObservableProperty] ObservableRangeCollection<string> payees = [];
+    [ObservableProperty] ObservableCollection<string> payees = [];
     [ObservableProperty] string findPayee;
     [ObservableProperty] string replacePayee;
     [ObservableProperty] string info;
@@ -38,6 +38,8 @@ public partial class BulkEditingViewModel : CommunityToolkit.Mvvm.ComponentModel
     List<Transaction> FoundTransactions { get; set; } = [];
     int TotalTransactionCount;
     int FoundTransactionCount;
+
+    StringBuilder infoSb = new();
 
     public BulkEditingViewModel(IAccountService _accountService, ICategoryService _categoryService,
         ITransactionService _transactionService, ILoggerService<BulkEditingViewModel> _logger)
@@ -74,7 +76,9 @@ public partial class BulkEditingViewModel : CommunityToolkit.Mvvm.ComponentModel
         try
         {
             var accounts = await accountService.GetOrderedAccounts();
-            Accounts.ReplaceRange(accounts);
+            Accounts.Clear();
+            foreach (var account in accounts)
+                Accounts.Add(account);
         }
         catch (Exception ex)
         {
@@ -91,7 +95,9 @@ public partial class BulkEditingViewModel : CommunityToolkit.Mvvm.ComponentModel
         try
         {
             var categories = await categoryService.GetParentCategories();
-            Categories.ReplaceRange(categories);
+            Categories.Clear();
+            foreach (var category in categories)
+                Categories.Add(category);
         }
         catch (Exception ex)
         {
@@ -141,7 +147,8 @@ public partial class BulkEditingViewModel : CommunityToolkit.Mvvm.ComponentModel
             }
 
             var subcategories = await categoryService.GetSubcategories(FindCategory);
-            FindSubcategories.AddRange(subcategories);
+            foreach (var subcategory in subcategories)
+                FindSubcategories.Add(subcategory);
         }
         catch (Exception ex)
         {
@@ -173,7 +180,8 @@ public partial class BulkEditingViewModel : CommunityToolkit.Mvvm.ComponentModel
             }
 
             var subcategories = await categoryService.GetSubcategories(ReplaceCategory);
-            ReplaceSubcategories.AddRange(subcategories);
+            foreach (var subcategory in subcategories)
+                ReplaceSubcategories.Add(subcategory);
         }
         catch (Exception ex)
         {
@@ -262,12 +270,12 @@ public partial class BulkEditingViewModel : CommunityToolkit.Mvvm.ComponentModel
 
     void UpdateText(int replacedCount = 0, bool found = false, bool replaced = false)
     {
-        StringBuilder sb = new();
-        sb.Append($"Total Transactions: {TotalTransactionCount}\n");
+        infoSb.Clear();
+        infoSb.Append($"Total Transactions: {TotalTransactionCount}\n");
         if (found)
-            sb.Append($"Transactions Found: {FoundTransactionCount}\n");
+            infoSb.Append($"Transactions Found: {FoundTransactionCount}\n");
         if (replaced)
-            sb.Append($"Transactions Replaced: {replacedCount} \n");
-        Info = sb.ToString();
+            infoSb.Append($"Transactions Replaced: {replacedCount} \n");
+        Info = infoSb.ToString();
     }
 }